reworking vercel api; reported bug to vercel
