+++ packages/frontend/hooks/api/messages/queries.ts
@@ -1,13 +1,18 @@
-import fetcher from "@/utils/fetcher";
-
 export const fetchOrbisMessages = async (
   orbisTag: string,
   { pageParam = 0 }: { pageParam?: number } = {}
 ) => {
   try {
-    const apiRoute: string = `/api/messages/${orbisTag}/${pageParam}`;
-    console.log(apiRoute);
-    return await fetcher(apiRoute);
+    const apiRoute: string = `/api/messages/orbisPosts`;
+    const response: Response = await fetch(apiRoute, {
+      headers: {
+        Accept: "application/json",
+        "Content-Type": "application/json",
+      },
+      method: "POST",
+      body: JSON.stringify({ postId: orbisTag, page: pageParam }),
+    });
+    return await response.json();
   } catch (error) {
     console.error("Error fetching messages", error);
     throw error;
